0.06 - time tracking, noLoad test
Removed doctype checking… all it really did was support the closing
slash on CSS files.

Added request completion time tracking… potentially useful to identify
which files are cached.

Added noLoad function/test to allow arbitrary test for the presence of
the library, even if it was loaded outside using(). Not sure about the
implementation, here.